cleanup and test
