+++ packages/meso-js/src/frameStore.ts
@@ -1,0 +1,3 @@
+import { FrameStore } from "./types";
+
+export const frameStore: FrameStore = {};
