+++ packages/meso-js/src/index.ts
@@ -2,6 +2,7 @@ import { version } from "../package.json";
 
 export * from "./types";
 export { transfer } from "./transfer";
+export { inlineTransfer } from "./inlineTransfer";
 export { validateLayout } from "./validateLayout";
 export { createPostMessageBus } from "./createPostMessageBus";
 export const MESO_JS_VERSION = version;
