--- packages/meso-js/src/store.ts
@@ -1,3 +1,0 @@
-import { Store } from "./types";
-
-export const store: Store = {};
