+++ packages/meso-js/src/utils.ts
@@ -1,0 +1,12 @@
+import { Environment } from "./types";
+
+export const apiHosts: { readonly [key in Environment]: string } = {
+  [Environment.LOCAL]: "http://localhost:5173",
+  [Environment.LOCAL_PROXY]: "http://localhost:4001",
+  [Environment.DEV]: "https://api.dev.meso.plumbing",
+  [Environment.PREVIEW]: "https://api.preview.meso.plumbing",
+  [Environment.SANDBOX]: "https://api.sandbox.meso.network",
+  [Environment.PRODUCTION]: "https://api.meso.network",
+};
+
+export const NOOP_TRANSFER_INSTANCE = { destroy: () => {} };
