+++ packages/meso-js/test/bus.test.ts
@@ -64,7 +64,6 @@ describe("setupBus", () => {
         frame,
         onEvent: onEventMock,
         onSignMessageRequest: onSignMessageRequestMock,
-        store: {},
       }),
     ).toThrowErrorMatchingInlineSnapshot(
       '"Unable to initialize @meso-network/meso-js. Invalid iframe configuration."',
@@ -80,7 +79,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(
       MessageKind.REQUEST_SIGNED_MESSAGE,
@@ -171,7 +169,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(MessageKind.CLOSE, expect.any(Function));
     const onCloseCallback = onMock.mock.calls.find(
@@ -205,7 +202,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(
       MessageKind.TRANSFER_UPDATE,
@@ -248,7 +244,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(
       MessageKind.TRANSFER_UPDATE,
@@ -288,7 +283,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(MessageKind.ERROR, expect.any(Function));
     const onErrorCallback = onMock.mock.calls.find(
@@ -318,7 +312,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(
       MessageKind.CONFIGURATION_ERROR,
@@ -354,7 +347,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(
       MessageKind.UNSUPPORTED_NETWORK_ERROR,
@@ -391,7 +383,6 @@ describe("setupBus", () => {
       frame,
       onEvent: onEventMock,
       onSignMessageRequest: onSignMessageRequestMock,
-      store: {},
     });
     expect(onMock).toBeCalledWith(
       MessageKind.UNSUPPORTED_ASSET_ERROR,
